switched comment routes to own file
